Refactor semantic module structure for better cohesion
- Consolidate `control_flow.rs` and `declarations.rs` into `src/semantic/statements.rs`.
- Keep `patterns.rs` separate to maintain acyclic dependency graph (`statements` -> `conversion` -> `patterns`).
- Remove obsolete modules.
- Ensure all tests pass.
